requirements.txt
git+git://github.com/Pycord-Development/pycord
aiohttp
python-dotenv